Fixes #22: Don't destroy paginated message when not active

+++ kord-extensions/src/main/kotlin/com/kotlindiscord/kord/extensions/pagination/Paginator.kt
@@ -297,6 +297,10 @@ public open class Paginator(
      * This will stop the paginator from processing events, and delete its message if [keepEmbed] is `false`.
      */
     public open suspend fun destroy(message: MessageBehavior) {
+        if (!active) {
+            return
+        }
+
         if (!keepEmbed) {
             message.delete()
         } else {
